Refactor unit tests entirely
Refactor unit tests to a tiny utility instead of True. The project requires
error testing which True can not handle. It is replaces with a couple of
very simple utility functions that handle Sass and expected CSS files.

Refactor all the tests to the new architecture.
Remove tests for the deprecated map format.
Remove True development dependency.